# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/movie": {
            "get": {
                "description": "get movies",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Get a list of movies",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order by (name, recent, last_scanned)",
                        "name": "orderby",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Title filter",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Alpha filter",
                        "name": "alpha",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Genre ID filter",
                        "name": "genre",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Crew person ID filter",
                        "name": "crew",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Person (cast/crew) ID filter",
                        "name": "person",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Cast person ID filter",
                        "name": "cast",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Country code filter",
                        "name": "country",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Show (multiple, unrated, duplicate, notitle, nodesc, watchlist)",
                        "name": "show",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/movielite.Movielist"
                        }
                    }
                }
            },
            "post": {
                "description": "create a new movie entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Create a movie",
                "parameters": [
                    {
                        "description": "Movie object",
                        "name": "movie",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Movie"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Movie"
                        }
                    }
                }
            }
        },
        "/api/movie/{id}": {
            "get": {
                "description": "get a movie by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Get a single movie",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Movie ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Movie"
                        }
                    }
                }
            },
            "put": {
                "description": "update a movie by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Update a movie",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Movie ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Movie object",
                        "name": "movie",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Movie"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Movie"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete a movie by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Delete a movie",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Movie ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Movie"
                        }
                    }
                }
            }
        },
        "/api/movie/{id}/addMeta/{metaid}": {
            "put": {
                "description": "Associates TMDB metadata with a movie",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Add/Update metadata for a movie",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Movie ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "TMDB Movie ID",
                        "name": "metaid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Movie"
                        }
                    }
                }
            }
        },
        "/api/movie/{id}/play": {
            "put": {
                "description": "play a movie file on the server",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Play a movie",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Movie ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Show directory in file manager instead of playing",
                        "name": "showdir",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "models.Credits": {
            "type": "object",
            "properties": {
                "cast": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Cast"
                    }
                },
                "crew": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Crew"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "tmdbmovieID": {
                    "type": "integer"
                }
            }
        },
...
        "models.File": {
            "type": "object",
            "properties": {
                "FileName": {
                    "type": "string"
                },
                "FullPath": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.Genres": {
            "type": "object",
            "properties": {
                "ID": {
                    "description": "ID      int64 `gorm:\"primary_key\"`",
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "tmdbmovieID": {
                    "type": "integer"
                }
            }
        },
        "models.Movie": {
            "type": "object",
            "required": [
                "File",
                "title"
            ],
            "properties": {
                "File": {
                    "$ref": "#/definitions/models.File"
                },
                "createdAt": {
                    "type": "string"
                },
                "deleted": {
                    "type": "boolean"
                },
                "deletedAt": {
                    "type": "string"
                },
                "file_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "is_tv": {
                    "type": "boolean"
                },
                "lastScanned": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/models.TMDBMovie"
                },
                "movieSearchResultsID": {
                    "type": "integer"
                },
                "multiplechoice": {
                    "$ref": "#/definitions/models.MovieSearchResults"
                },
                "org_name": {
                    "type": "string"
                },
                "rating": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "tmdbmovieID": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "watchlist": {
                    "type": "boolean"
                }
            }
        },
        "models.MovieSearchResults": {
            "type": "object",
            "properties": {
                "Results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MovieShort"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                },
                "total_results": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.MovieShort": {
            "type": "object",
            "properties": {
                "adult": {
                    "type": "boolean"
                },
                "backdrop_path": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "movieSearchResultsID": {
                    "type": "integer"
                },
                "original_title": {
                    "description": "ID                   int    `json:\"id\"`",
                    "type": "string"
                },
                "overview": {
                    "type": "string"
                },
                "popularity": {
                    "description": "GenreIDs             []int32 `json:\"genre_ids\"`",
                    "type": "number"
                },
                "poster_path": {
                    "type": "string"
                },
                "release_date": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "video": {
                    "type": "boolean"
                },
                "vote_average": {
                    "type": "number"
                },
                "vote_count": {
                    "type": "integer"
                }
            }
        },
        "models.ProductionCompanies": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "logo_path": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "origin_country": {
                    "type": "string"
                }
            }
        },
        "models.ProductionCountries": {
            "type": "object",
            "properties": {
                "iso_3166_1": {
                    "description": "TMDBMovieMovieID uint",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.SpokenLanguages": {
            "type": "object",
            "properties": {
                "iso_639_1": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.TMDBMovie": {
            "type": "object",
            "properties": {
                "Genres": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Genres"
                    }
                },
                "adult": {
                    "description": "TMDBMovieID int        `json:\"ID\"`",
                    "type": "boolean"
                },
                "backdrop_path": {
                    "description": "MovieID      int64  `gorm:\"primary_key\" json:\"id\"`",
                    "type": "string"
                },
                "budget": {
                    "description": "BelongsToCollection bool   `json:\"belongs_to_collection\"`\nBelongsToCollection CollectionShort `json:\"belongs_to_collection\"`",
                    "type": "integer",
                    "format": "int32"
                },
                "createdAt": {
                    "type": "string"
                },
                "credits": {
                    "description": "AlternativeTitles *MovieAlternativeTitles `json:\"alternative_titles,omitempty\"`",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Credits"
                        }
                    ]
                },
                "deletedAt": {
                    "type": "string"
                },
                "homepage": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "imdb_id": {
                    "description": "ID               int",
                    "type": "string"
                },
                "original_language": {
                    "type": "string"
                },
                "original_title": {
                    "type": "string"
                },
                "overview": {
                    "type": "string"
                },
                "popularity": {
                    "type": "number",
                    "format": "float32"
                },
                "poster_path": {
                    "type": "string"
                },
                "production_companies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ProductionCompanies"
                    }
                },
                "production_countries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ProductionCountries"
                    }
                },
                "release_date": {
                    "type": "string"
                },
                "revenue": {
                    "type": "integer",
                    "format": "int32"
                },
                "runtime": {
                    "type": "integer",
                    "format": "int32"
                },
                "spoken_languages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SpokenLanguages"
                    }
                },
                "status": {
                    "type": "string"
                },
                "tagline": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "video": {
                    "type": "boolean"
                },
                "vote_average": {
                    "type": "number"
                },
                "vote_count": {
                    "type": "integer"
                }
            }
        },
...
        "movielite.Movielist": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Movie"
                    }
                },
                "meta": {
                    "$ref": "#/definitions/movielite.meta"
                }
            }
        },
...
        "movielite.meta": {
            "type": "object",
            "properties": {
                "total": {
                    "type": "integer"
                }
            }
        },
...
    }
}